Simplify deposit submission builder
